+++ src/core/io/lottie/lottie_html_format.cpp
@@ -2,18 +2,18 @@
 #include "lottie_format.hpp"
 #include "cbor_write_json.hpp"
 
-bool io::lottie::LottieHtmlFormat::on_save(QIODevice& file, const QString&,
-                                           model::Document* document, const QVariantMap&)
+QByteArray io::lottie::LottieHtmlFormat::html_head(ImportExport* ie, model::Document* document, const QString& extra)
 {
-    file.write(QString(
+    return QString(
 R"(<!DOCTYPE html>
 <html>
 <head>
     <meta charset="utf-8" />
+    <title>%4: %5</title>
     <style>
         html, body { width: 100%; height: 100%; margin: 0; }
         body { display: flex; }
-        #bodymovin { width: %1px; height: %2px; margin: auto;
+        #animation { width: %1px; height: %2px; margin: auto;
             background-color: white;
             background-size: 64px 64px;
             background-image:
@@ -23,23 +23,34 @@ R"(<!DOCTYPE html>
                 linear-gradient(to right, transparent 50%, rgba(0, 0, 0, .5) 50%);
         }
     </style>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.5.3/lottie.js"></script>
+    %3
 </head>
+)")
+    .arg(document->main()->width.get())
+    .arg(document->main()->height.get())
+    .arg(extra)
+    .arg(document->main()->object_name())
+    .arg(ie->name())
+    .toUtf8()
+    ;
+}
+
+bool io::lottie::LottieHtmlFormat::on_save(QIODevice& file, const QString&,
+                                           model::Document* document, const QVariantMap&)
+{
+    file.write(html_head(this, document, "<script src='https://cdnjs.cloudflare.com/ajax/libs/bodymovin/5.5.3/lottie.js'></script>"));
+    file.write(R"(
 <body>
-<div id="bodymovin"></div>
+<div id="animation"></div>
 
 <script>
     var animData = {
-        container: document.getElementById('bodymovin'),
+        container: document.getElementById('animation'),
         renderer: 'svg',
         loop: true,
         autoplay: true,
         animationData:
-)")
-    .arg(document->main()->width.get())
-    .arg(document->main()->height.get())
-    .toUtf8()
-    );
+)");
 
     file.write(cbor_write_json(LottieFormat().to_json(document), false));